big edit to get work well lgsm.sh, delete useless repetition add some conditions.

+++ lgsm.sh
@@ -9,12 +9,25 @@ ServerType="arkserver"
 ## Image name to run (i have build with the lgsm-build.sh)
 Img="lgsm-docker"
 
-if [ $1 != null ]
+## check if the container already running (true or '')
+status=$(sudo docker inspect --format="{{.State.Running}}" $InstanceName 2> /dev/null)
+
+if [ "$status" != "true" ] && [ $1 != "stop" ]
 then
-        echo $"Usage: $0 {start|stop|restart|console|monitor|update|backup|cronjob}"
-	cmd=$@
+	echo "docker container was not running. start it for you."
+	sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash 2> /dev/null
+elif [ "$status" == "true" ]
+then
+	echo "docker container already running, append command."
 else
-	echo $"Usage: $0 {start|stop|restart|console|monitor|update|backup|cronjob}"
+	echo "docker container not running."
+fi
+
+if [ "$1" != "" ]
+then
+       	cmd=$1
+else
+	echo $"Usage: $0 {start|stop|restart|console|monitor|update|backup|attach|command|cronjob}"
 	read -a cmd
 fi
 
@@ -22,55 +35,58 @@ fi
 case $cmd in
         "install")
             read -a type
-	    sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash 2> /dev/null
 	    sudo docker exec $InstanceName $ServerType install $type
             ;;
 
         "start")
-	    sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash 2> /dev/null
             sudo docker exec $InstanceName $ServerType start
             ;;
 
         "stop")
-	    sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash 2> /dev/null
-            sudo docker exec $InstanceName $ServerType stop
-	    sudo docker kill $InstanceName
+	    if [ "$status" == "true" ]
+	    then
+            	sudo docker exec $InstanceName $ServerType stop
+	    	sudo docker kill $InstanceName
+	    fi
             ;;
 
         "restart")
-	    sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash 2> /dev/null
             sudo docker exec $InstanceName $ServerType restart
             ;;
 
         "console")
-	    sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash 2> /dev/null
             sudo docker exec $InstanceName $ServerType console
             ;;
 
         "monitor")
-	    sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash 2> /dev/null
             sudo docker exec $InstanceName $ServerType monitor
             ;;
 
-        "update") ## update stop the server if is already running.
-	    sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash
+        "update") ## update stop the server if is already running(lgsm script).
             sudo docker exec $InstanceName $ServerType update
             ;;
 
         "backup")
-	    sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash
             sudo docker exec $InstanceName $ServerType backup
             ;;
 
         "conjob") ## need to be test.
-            sudo docker run --name $InstanceName --rm -it -d -v "/home/lgsm/:/home/lgsm" $Img bash
             sudo docker exec $InstanceName crontab -l > tempcronfile
 	    sudo docker exec $InstanceName echo "0 5 * * *  /home/lgsm/$ServerType update > /dev/null 2>&1" > tempcronfile
  	    sudo docker exec $InstanceName crontab tempcronfile && rm tempcronfile
             ;;
 
+	"attach")
+	    sudo docker attach $InstanceName
+	    ;;
+
+	"command")
+	    echo "input the complete command to execute in virtual container"
+	    read -a $command
+	    sudo docker exec $InstanceName $command
+	    ;;
+
         *)
-            echo $"Usage: $0 {start|stop|restart|console|monitor|update|backup}"
             exit 1
 
 esac
